refactor(frontend): dynamic view layer state
Configure view layers in `config/view-layers`, then load the layer state dynamically in `state/layers/view-layers`. Move dynamic modules into `state/layers/modules` since Rollup doesn't allow dynamic imports from `./`.
