+++ frontend/src/config/view-layers.ts
@@ -1,0 +1,12 @@
+export const VIEW_LAYERS = [
+  'regions',
+  'droughtRegions',
+  'terrestrial',
+  'marine',
+  'hazards',
+  'buildings',
+  'networks',
+  'droughtOptions',
+  'featureBoundingBox',
+  'labels',
+] as string[];
