+++ frontend/src/state/layers/view-layers-params.ts
@@ -5,7 +5,7 @@ import { StyleParams, ViewLayer, ViewLayerParams } from 'lib/data-map/view-layer
 import { viewLayersFlatState } from './view-layers-flat';
 import _ from 'lodash';
 import { selectionState } from 'state/interactions/interaction-state';
-import { networkStyleParamsState } from './networks';
+import { networkStyleParamsState } from './modules/networks';
 
 export const viewLayerState = atomFamily<ViewLayer, string>({
   key: 'viewLayerState',
